--- phoenix-hadoop2-compat/src/main/resources/META-INF/services/org.apache.phoenix.metrics.MetricsManager
@@ -1,1 +1,0 @@
-org.apache.phoenix.metrics.MetricsManagerImpl
