--- phoenix-hadoop2-compat/src/main/resources/META-INF/services/org.apache.phoenix.trace.TestableMetricsWriter
@@ -1,1 +1,0 @@
-org.apache.phoenix.trace.PhoenixMetricsSink
