--- phoenix-hadoop2-compat/src/test/java/org/apache/phoenix/trace/TracingTest.java
@@ -1,34 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.phoenix.trace;
-
-import static org.junit.Assert.assertNotNull;
-
-import org.junit.Test;
-
-public class TracingTest {
-
-    /**
-     * Test that we can correctly load a class that will convert the tracing output to metrics
-     * @throws Exception on failure
-     */
-    @Test
-    public void testLoadTracingToMetrics() throws Exception{
-        assertNotNull("Didn't find a trace receiver", TracingCompat.newTraceMetricSource());
-    }
-}
